Study2-EmoGist/requirements.txt
aiofiles==23.2.1
annotated-types==0.6.0
anyio==4.3.0
appnope==0.1.4
asgiref==3.7.2
asttokens==2.4.1
Authlib==0.14.3
beautifulsoup4==4.12.3
cachetools==4.2.4
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.1
cryptography==42.0.5
debugpy==1.8.1
decorator==5.1.1
emoji==2.11.0
executing==2.0.1
fastapi==0.68.0
google==3.0.0
google-api-core==2.10.2
google-api-python-client==2.15.0
google-auth==1.32.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==0.4.4
googleapis-common-protos==1.62.0
h11==0.14.0
httpcore==1.0.4
httplib2==0.22.0
httpx==0.27.0
idna==3.6
ipykernel==6.29.3
ipython==8.22.2
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.0.1
jose==1.0.0
jupyter_client==8.6.0
jupyter_core==5.7.1
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
nest-asyncio==1.6.0
numpy==1.26.2
oauthlib==3.2.2
openai==1.17.0
opencv-python==4.8.1.78
packaging==23.2
parso==0.8.3
pexpect==4.9.0
pip==24.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.14
pydantic_core==2.16.3
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pyzmq==25.1.2
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
setuptools==65.5.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
stack-data==0.6.3
starlette==0.14.2
tornado==6.4
traitlets==5.14.1
typing_extensions==4.10.0
uritemplate==3.0.1
urllib3==2.2.1
uvicorn==0.15.0
wcwidth==0.2.13
oauth2client==4.1.3
